requirements.txt
aiobotocore
fastapi[all]
python-multipart
pymongo[srv]
pyjwt
passlib[bcrypt]